split store and api in separated files
