+++ src/index.js
@@ -1,23 +1,16 @@
-import path from 'path';
 import {
   assign,
-  find,
-  get as get_,
-  has,
-  identity,
-  keys,
-  pick,
-  startsWith
+  pick
 } from 'lodash/fp';
 import Etcd from 'node-etcd';
 import {Observable} from 'rxjs';
 import createDebug from 'debug';
 
 import {createEtcd$} from './etcd';
-import createCombiner$ from './combiner';
 import createFallback$ from './fallback';
-import createSave from './save';
-import createIndexer from './indexer';
+import createCombiner$ from './combiner';
+import createStore from './store';
+import createAPI from './api';
 
 const debug = createDebug('squirrel');
 
@@ -38,70 +31,18 @@ const createSquirrel = options => {
 
   const client = new Etcd(options.hosts, pick(['auth', 'ca', 'key', 'cert'], options));
   const watcher = client.watcher(options.cwd, null, {recursive: true});
-  const save = options.save ? createSave(options.fallback) : identity;
-  const indexer = createIndexer(options.indexes);
 
   const events$ = Observable.concat(
     createFallback$(options.fallback),
     createEtcd$(client, watcher, options.cwd)
   );
 
-  const store$ = save(createCombiner$(events$)).map(node => ({
-    node,
-    indexes: indexer(node)
-  }));
-
-  let store = null;
-  const storeReady = store$.take(1).do(_store => {
-    store = _store;
-  }).toPromise();
-  const subscription = store$.subscribe(_store => {
-    store = _store;
-  });
-
-  subscription.add(() => {
-    debug('Unsubscribe');
-    watcher.stop();
-  });
-
-  const ready = key => storeReady.then(() => store[key]);
-
-  const getBy = (index, key) => {
-    debug(`getBy: ${index} => ${key}`);
-    return ready('indexes').then(indexes => {
-      return has(key, indexes[index]) ?
-      get_(key, indexes[index]).value :
-      null;
-    });
-  };
-
-  const getAll = index => {
-    debug(`getAll => ${index}`);
-    return ready('indexes').then(indexes => {
-      return keys(indexes && indexes[index]);
-    });
-  };
-
-  const get = path => {
-    debug(`get => ${path}`);
-    return ready('node').then(node => {
-      return _get(path, node);
-    });
-  };
-
-  const _get = (_path, node) => {
-    if (!node) return null;
-    if (path.relative(node.key, _path) === '') return node;
-
-    return _get(_path, find(function(child) {
-      return startsWith(child.key, _path);
-    }, node.nodes));
-  };
+  const combiner$ = createCombiner$(events$);
+  const {store, subscription} = createStore(options, combiner$, watcher);
+  const api = createAPI(store);
 
   return {
-    getBy,
-    getAll,
-    get,
+    ...api,
     close: () => subscription.unsubscribe()
   };
 };
